requirements.txt
annotated-types==0.7.0
anyio==4.12.1
black==25.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
wheel==0.45.1
